backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
ollama>=0.6.1
qdrant-client>=1.11,<1.13
redis>=5.0
psycopg2-binary>=2.9
sqlalchemy>=2.0
alembic>=1.13
python-multipart>=0.0.9
pydantic>=2.0
pydantic-settings>=2.0
httpx>=0.27
